Add example greedy search using associated moves
Add example greedy search using associated moves for the TSP problem.
The number of cities and the random seed can be adjusted using the
sliders in the `TSP-demo.ipynb` notebook.  The search can be run by
executing the remaining cells in the notebook, after which the cost at
each iteration will be plotted.